splitted up config and server
